Move to a new line test after full stop.
The intention of this is reduce the lines edited in the future, due to column with of 80 chars we have.
giving some room to new text to be added without having to modify the whole paragraph that could be of many lines

Example:
It's our hope that this will become a central hub for community discussion on
best practices in Elixir. Feel free to open tickets or send pull requests with
improvements. Thanks in advance for your help!

It's our hope that this will become a central hub for community discussion on
best practices in Elixir.
Feel free to open tickets or send pull requests with improvements.
Thanks in advance for your help!

In the first example if you want change the first sentence and it will take 10 chars more,
this space will be translated to the following sentences having to modify all subsequent lines.

In the second eample if you want to add text and that text goes beyond the
80chars it will not be translated to the following sentences.

ADDITIONALLY:
I have added an empty line between items in the lists.
It looks better and it is also consistent with the rest of the guide which pretty much has
spaces becouse of the code-blocks.