Add parentheses preference for anonymous function arguments
At this point, the style guide has an item indicating a preference for
inclusion of parentheses when there are arguments to a named function,
but it does not mention such a preference for anonymous functions. This
change adds a preference for the inclusion of parentheses when
arguments exist for anonymous functions, that matches the existing
preference for named functions. A concession is made for the case when
there is a single argument to an anonymous function.