Name do_ functions if name and arity both match
Why
---

The convention is to add `do_` to the name only if one is forced by a
conflict in both the name and arity. This is because functions are
identified by the combination of name and arity in Elixir/Erlang. If
the name is the same yet the arity is different, it is a different
function.

How
---

Update `#private-functions-with-same-name-as-public` rule